+++ src/__test__/export.test.ts
@@ -21,6 +21,17 @@ const {
 	Subscribe,
 	SubscribeForm,
 	Social,
+	SocialMediaBar,
+	FacebookLink,
+	FacebookShare,
+	TwitterLink,
+	TwitterShare,
+	LinkedInLink,
+	LinkedInShare,
+	EmailShare,
+	InstagramLink,
+	GithubLink,
+	RSSLink,
 	Footer,
 	Latest,
 	Recent,
@@ -70,6 +81,7 @@ const {
 	getQueryString,
 	titleCase,
 	isSubscriberCookieSet,
+	getSubscriberId,
 	subscribe,
 	useContent,
 	useEntries,
