+++ src/components/Footer/__test__/Footer.test.tsx
@@ -1,6 +1,14 @@
 import renderer from 'react-test-renderer';
 import Footer from '..';
-import { Bar, FacebookLink, InstagramLink, TwitterLink, GithubLink, LinkedInLink, RSSLink } from '../../SocialMedia';
+import {
+	SocialMediaBar,
+	FacebookLink,
+	InstagramLink,
+	TwitterLink,
+	GithubLink,
+	LinkedInLink,
+	RSSLink,
+} from '../../SocialMedia';
 import Copyright from '../../Copyright';
 import Logo from '../../Logo';
 import Subscribe from '../../Subscribe';
@@ -12,14 +20,14 @@ test('Test complete', () => {
 	const component = renderer.create(
 		<Footer
 			social={
-				<Bar>
+				<SocialMediaBar>
 					<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 					<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 					<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 					<GithubLink href="https://github.com/pinpt" newTab />
 					<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 					<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-				</Bar>
+				</SocialMediaBar>
 			}
 			copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 			subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -52,14 +60,14 @@ test('Test social only', () => {
 	const component = renderer.create(
 		<Footer
 			social={
-				<Bar>
+				<SocialMediaBar>
 					<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 					<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 					<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 					<GithubLink href="https://github.com/pinpt" newTab />
 					<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 					<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-				</Bar>
+				</SocialMediaBar>
 			}
 		/>
 	);
