+++ src/components/Page/__stories__/PageEntry.stories.tsx
@@ -10,7 +10,7 @@ import Tags from '../../Tags';
 import Clap from '../../Clap';
 import Footer from '../../Footer';
 import {
-	Bar,
+	SocialMediaBar,
 	FacebookLink,
 	FacebookShare,
 	InstagramLink,
@@ -70,26 +70,26 @@ export const Default: React.VFC<{}> = () => {
 					tags={<Tags.Bar tags={['feature', 'improvement', 'mobile']} />}
 					clap={<Clap clapCount={count} handleClap={() => setCount((c) => c + 1)} />}
 					sharing={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookShare href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<TwitterShare href="https://twitter.com/pinpoint_sw" newTab />
 							<LinkedInShare href="https://linkedin.com/company/pinpoint-software" newTab />
 							<EmailShare recipient="hello@pinpoint.com" />
-						</Bar>
+						</SocialMediaBar>
 					}
 				/>
 			}
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -119,26 +119,26 @@ export const With_Pagination: React.VFC<{}> = () => {
 					tags={<Tags.Bar tags={['feature', 'improvement', 'mobile']} />}
 					clap={<Clap clapCount={count} handleClap={() => setCount((c) => c + 1)} />}
 					sharing={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookShare href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<TwitterShare href="https://twitter.com/pinpoint_sw" newTab />
 							<LinkedInShare href="https://linkedin.com/company/pinpoint-software" newTab />
 							<EmailShare recipient="hello@pinpoint.com" />
-						</Bar>
+						</SocialMediaBar>
 					}
 				/>
 			}
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -176,26 +176,26 @@ export const Not_Zoomable: React.VFC<{}> = () => {
 					tags={<Tags.Bar tags={['feature', 'improvement', 'mobile']} />}
 					clap={<Clap clapCount={count} handleClap={() => setCount((c) => c + 1)} />}
 					sharing={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookShare href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<TwitterShare href="https://twitter.com/pinpoint_sw" newTab />
 							<LinkedInShare href="https://linkedin.com/company/pinpoint-software" newTab />
 							<EmailShare recipient="hello@pinpoint.com" />
-						</Bar>
+						</SocialMediaBar>
 					}
 				/>
 			}
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
