+++ src/components/Page/__test__/PageEntry.test.tsx
@@ -7,7 +7,15 @@ import Header from '../../Header';
 import Logo from '../../Logo';
 import { Document } from '../../Renderer';
 import Sidebar from '../../Sidebar';
-import { Bar, FacebookLink, InstagramLink, TwitterLink, GithubLink, LinkedInLink, RSSLink } from '../../SocialMedia';
+import {
+	SocialMediaBar,
+	FacebookLink,
+	InstagramLink,
+	TwitterLink,
+	GithubLink,
+	LinkedInLink,
+	RSSLink,
+} from '../../SocialMedia';
 import Subscribe from '../../Subscribe';
 import Tags from '../../Tags';
 import Page from '../Entry';
@@ -45,14 +53,14 @@ test('Test full entry', () => {
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -104,14 +112,14 @@ test('Test no sidebar', () => {
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -146,14 +154,14 @@ test('Test with pagination', () => {
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
@@ -196,14 +204,14 @@ test('Test not zoomable entry', () => {
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
