+++ src/components/Page/__test__/PageError.test.tsx
@@ -3,7 +3,15 @@ import Copyright from '../../Copyright';
 import Error from '../../Error';
 import Footer from '../../Footer';
 import Logo from '../../Logo';
-import { Bar, FacebookLink, InstagramLink, TwitterLink, GithubLink, LinkedInLink, RSSLink } from '../../SocialMedia';
+import {
+	SocialMediaBar,
+	FacebookLink,
+	InstagramLink,
+	TwitterLink,
+	GithubLink,
+	LinkedInLink,
+	RSSLink,
+} from '../../SocialMedia';
 import Subscribe from '../../Subscribe';
 import Page from '../Error';
 
@@ -17,14 +25,14 @@ test('Test full error', () => {
 			footer={
 				<Footer
 					social={
-						<Bar>
+						<SocialMediaBar>
 							<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 							<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 							<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 							<GithubLink href="https://github.com/pinpt" newTab />
 							<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 							<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-						</Bar>
+						</SocialMediaBar>
 					}
 					copyright={<Copyright text="2021 by Pinpoint Software, Inc." logo={<Logo src={IMAGE_URL} />} />}
 					subscribe={<Subscribe href="https://pinpoint.com" />}
