+++ src/components/SocialMedia/__test__/SocialBar.test.tsx
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import Bar from '../Bar';
+import SocialMediaBar from '../SocialMediaBar';
 import FacebookLink from '../Facebook/FacebookLink';
 import InstagramLink from '../Instagram/InstagramLink';
 import TwitterLink from '../Twitter/TwitterLink';
@@ -9,14 +9,14 @@ import RSSLink from '../RSS/RSSLink';
 
 test('Test all', () => {
 	const component = renderer.create(
-		<Bar>
+		<SocialMediaBar>
 			<FacebookLink href="https://www.facebook.com/Pinpoint.Engineering" newTab />
 			<InstagramLink href="https://www.instagram.com/pinpoint_sw/" newTab />
 			<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 			<GithubLink href="https://github.com/pinpt" newTab />
 			<LinkedInLink href="https://linkedin.com/company/pinpoint-software" newTab />
 			<RSSLink href="https://api.pinpoint.com.so/rss/PirxVTE94u3YmGNOySRY" newTab />
-		</Bar>
+		</SocialMediaBar>
 	);
 	const tree = component.toJSON();
 	expect(tree).toMatchSnapshot();
@@ -24,10 +24,10 @@ test('Test all', () => {
 
 test('Test short', () => {
 	const component = renderer.create(
-		<Bar>
+		<SocialMediaBar>
 			<TwitterLink href="https://twitter.com/pinpoint_sw" newTab />
 			<GithubLink href="https://github.com/pinpt" newTab />
-		</Bar>
+		</SocialMediaBar>
 	);
 	const tree = component.toJSON();
 	expect(tree).toMatchSnapshot();
@@ -35,9 +35,9 @@ test('Test short', () => {
 
 test('Test single', () => {
 	const component = renderer.create(
-		<Bar>
+		<SocialMediaBar>
 			<GithubLink href="https://github.com/pinpt" newTab />
-		</Bar>
+		</SocialMediaBar>
 	);
 	const tree = component.toJSON();
 	expect(tree).toMatchSnapshot();
@@ -45,9 +45,9 @@ test('Test single', () => {
 
 test('Custom className', () => {
 	const component = renderer.create(
-		<Bar className="custom-test">
+		<SocialMediaBar className="custom-test">
 			<GithubLink href="https://github.com/pinpt" newTab />
-		</Bar>
+		</SocialMediaBar>
 	);
 	const tree = component.toJSON();
 	expect(tree).toMatchSnapshot();
